readability index

+++ src/main/groovy/com/tbp/etl/model/Comment.groovy
@@ -38,7 +38,7 @@ class Comment {
     Integer period
 
     @Column(name = "ari")
-    Integer ari
+    Double ari
     @Column(name = "smog")
     Double smog
     @Column(name = "flesch_reading")
